store most state in jotai
