+++ app/layout.tsx
@@ -3,6 +3,8 @@ import "./globals.css";
 import { Analytics } from "@vercel/analytics/react";
 import { Footer } from "../components/Footer";
 import { Toaster } from "@/components/ui/toaster";
+import { ProjectDataProvider } from "@/components/ProjectDataProvider";
+import { Provider } from "jotai";
 
 export const metadata: Metadata = {
     title: "MSIT Master Proposals 2025",
@@ -19,10 +21,14 @@ export default function RootLayout({
     return (
         <html lang="en">
             <body>
-                <Analytics />
-                {children}
-                <Footer />
-                <Toaster />
+                <Provider>
+                    <ProjectDataProvider>
+                        <Analytics />
+                        {children}
+                        <Footer />
+                        <Toaster />
+                    </ProjectDataProvider>
+                </Provider>
             </body>
         </html>
     );
