+++ package.json
@@ -18,6 +18,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "dotenv": "^16.4.7",
+    "jotai": "^2.12.2",
     "jsdom": "^26.0.0",
     "lucide-react": "^0.484.0",
     "next": ">=15.2.3",
